OpenCV Fisheye wide-angle distortion correction For Android
At present the community rarely on the Android version of the OpenCV data, especially the distortion of wide-angle fisheye data is less, these days I wrote a fisheye wide-angle distortion correction of small items can be for reference and learning. Of course there are still a lot of problems in the hope that we can work together to improve. This project is developed using studio Android, which integrates the OpenCV3.0 library. I hope you all have a good time. Thank you